+++ lib/llm_cost_tracker/billing/components.rb
@@ -14,9 +14,7 @@ module LlmCostTracker
       "per_request" => 1,
       "per_1k_requests" => 1_000,
       "per_session" => 1,
-      "per_hour" => 1,
-      "per_gb_day" => 1,
-      "per_image" => 1
+      "per_hour" => 1
     }.freeze
 
     DEFAULT_RATE_BASIS_BY_UNIT = {
@@ -24,8 +22,7 @@ module LlmCostTracker
       "character" => "per_million_characters",
       "request" => "per_request",
       "session" => "per_session",
-      "hour" => "per_hour",
-      "image" => "per_image"
+      "hour" => "per_hour"
     }.freeze
 
     module Components